Extracted 'OnboadingOptions' Enum into it's own file for now.
